Fix warnings

+++ src/phantasm-hardware-interface/d3d12/cmd_list_translation.cc
@@ -742,8 +742,8 @@ void phi::d3d12::command_list_translator::execute(const phi::cmd::copy_texture_t
 
 #ifdef CC_ENABLE_ASSERTIONS
     auto const& srcDescFull = _globals.pool_resources->getTextureDescription(copy_text.source);
-    CC_ASSERT(sourceBox.right <= srcDescFull.width && sourceBox.bottom <= srcDescFull.height && sourceBox.back <= srcDescFull.depth_or_array_size
-              && "Source box out of bounds");
+    CC_ASSERT((int)sourceBox.right <= srcDescFull.width && (int)sourceBox.bottom <= srcDescFull.height
+              && (int)sourceBox.back <= srcDescFull.depth_or_array_size && "Source box out of bounds");
 #endif
 
     _cmd_list->CopyTextureRegion(&dest, 0, 0, 0, &source, &sourceBox);
